etc : Update

+++ handler/userHandler.go
@@ -1,7 +1,6 @@
 package handler
 
 import (
-	"fmt"
 	"net/http"
 	"strconv"
 
@@ -25,8 +24,17 @@ func GetUserList(c *gin.Context) {
 // DeleteUser func
 func DeleteUser(c *gin.Context) {
 	uID := c.Query("uID")
+	if uID == "" {
+		c.JSON(http.StatusOK, utils.JSONReturnMsg(
+			false, "올바른 uID 값을 입력해주세요",
+		))
+		return
+	}
 
 	mysql.DeleteUser(uID)
+	c.JSON(http.StatusOK, utils.JSONReturnMsg(
+		true, "삭제되었습니다",
+	))
 }
 
 // CreateUser func
@@ -35,10 +43,6 @@ func CreateUser(c *gin.Context) {
 	user.Uemail = c.PostForm("uEmail")
 	user.Upw = c.PostForm("uPW")
 	user.Uname = c.PostForm("uName")
-	if user.Uemail == "" || user.Upw == "" || user.Uname == "" {
-		c.String(http.StatusOK, "Uemail, Upw, Uname, uAgree 값은 필수 입력사항입니다.")
-		return
-	}
 	user.Ugender = c.DefaultPostForm("uGender", "null")
 	user.Uaddr = c.DefaultPostForm("uAddr", "null")
 	user.Upost, _ = strconv.Atoi(c.DefaultPostForm("uPost", "0"))
@@ -46,7 +50,18 @@ func CreateUser(c *gin.Context) {
 	user.Ubirth = c.DefaultPostForm("uBirth", "2000-01-01")
 	user.Uagree, _ = strconv.Atoi(c.DefaultPostForm("uAgree", "0"))
 
+	if user.Uemail == "" || user.Upw == "" || user.Uname == "" {
+		c.JSON(http.StatusOK, utils.JSONReturnMsg(
+			false, "Uemail, Upw, Uname, uAgree 값은 필수 입력사항입니다",
+		))
+		return
+	}
+
 	mysql.InsertNewUser(user)
+
+	c.JSON(http.StatusOK, utils.JSONReturnMsg(
+		true, "유저가 생성되었습니다",
+	))
 }
 
 // LoginUser func
@@ -56,23 +71,30 @@ func LoginUser(c *gin.Context) {
 	user.Upw = c.PostForm("uPW")
 
 	if user.Uemail == "" || user.Upw == "" {
-		c.String(http.StatusOK, "Uemail, Upw 값은 필수 입력사항입니다.")
+		c.JSON(http.StatusOK, utils.JSONReturnMsg(
+			false, "uEmail, uPW 값은 필수 입력사항입니다",
+		))
 		return
 	}
 	exist := mysql.CheckExist(user.Uemail)
-	fmt.Println(exist)
 	if exist {
 		status := mysql.LoginUser(user)
 
 		if status {
-			c.JSON(http.StatusOK, gin.H{"login": "Success"})
+			c.JSON(http.StatusOK, utils.JSONReturnMsg(
+				true, "로그인에 성공했습니다",
+			))
 			return
 		}
-		c.JSON(http.StatusOK, gin.H{"login": "Fail"})
+		c.JSON(http.StatusOK, utils.JSONReturnMsg(
+			false, "잘못된 비밀번호 입니다",
+		))
 		return
 	}
 
-	c.JSON(http.StatusOK, gin.H{"msg": "존재하지 않는 아이디 입니다"})
+	c.JSON(http.StatusOK, utils.JSONReturnMsg(
+		false, "존재하지 않는 아이디 입니다",
+	))
 	return
 
 }
@@ -83,10 +105,14 @@ func CheckExist(c *gin.Context) {
 	exist := mysql.CheckExist(uEmail)
 
 	if exist {
-		c.JSON(http.StatusOK, gin.H{"msg": "이미 존재하는 아이디 입니다."})
+		c.JSON(http.StatusOK, utils.JSONReturnMsg(
+			false, "이미 존재하는 아이디 입니다",
+		))
 		return
 	}
 
-	c.JSON(http.StatusOK, gin.H{"msg": "사용가능한 아이디 입니다."})
+	c.JSON(http.StatusOK, utils.JSONReturnMsg(
+		true, "사용가능한 아이디 입니다",
+	))
 	return
 }
